third_party/WebKit/LayoutTests/external/wpt/payment-request/payment-request-insecure.http.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test for PaymentRequest Constructor (insecure)</title>
<link rel="help" href="https://w3c.github.io/payment-request/#paymentrequest-interface">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_false(isSecureContext);
  assert_false('PaymentRequest' in window);
  assert_false('PaymentResponse' in window);
  assert_false('PaymentRequestUpdateEvent' in window);
}, "PaymentRequest constructor must not be exposed in insecure context");
</script>
